Add basic implementation in the console
userInput is written as a string then transformed into a number
from a choice Array. Then we get a random number between 1 and 3
from the computer. We compare those two and that gives us the result
of each round. We loop around the game logic to play the number of
round we want: 5. Then by decrementing when the player loses and
incrementing when the player wins (nothing if tie), we get who won
the Bo5 in the end.